Added POST Request
Added Axios POST to smurf forms, then clearing out values on SmurfForms state. Appjs state needed to be updated shortly after this, so passed a function from Appjs as a prop to update state. included a .5 second timeout before this occurs so that the ajax request is finished processing by the time we update Apps state.